+++ sdk/dotnet/GetRegion.cs
@@ -12,7 +12,7 @@ namespace Pulumi.Linode
     public static class GetRegion
     {
         /// <summary>
-        /// `linode..getRegion` provides details about a specific Linode region.
+        /// `linode..getRegion` provides details about a specific Linode region. See all regions [here](https://api.linode.com/v4/regions).
         /// 
         /// {{% examples %}}
         /// {{% /examples %}}